requirements.txt
 # Web scraping and HTTP requests
requests  #==2.31.0
beautifulsoup4 #==4.12.2
selenium #==4.15.2
lxml #==4.9.3

 # Database connectivity
pyodbc #==5.0.1

 # Data processing and analysis
pandas #==2.1.3
scikit-learn #==1.3.2
scipy #==1.11.4

 # Machine learning and NLP
nltk #==3.8.1
textblob #==0.17.1
wordcloud #==1.9.2

 # Utilities
python-dotenv #==1.0.0
schedule #==1.2.0
tqdm #==4.66.1

 # Web framework (optional for API)
flask #==3.0.0
flask-cors #==4.0.0

 # Image processing
Pillow #==10.1.0

 # Logging
colorlog #==6.8.0
